split azure constants to new file
