+++ src/client/src/components/ResourceGroupSelection/index.tsx
@@ -6,7 +6,7 @@ import { ReactComponent as Refresh } from "../../assets/i-refresh.svg";
 import messages from "./messages";
 import Dropdown from "../Dropdown";
 import classNames from "classnames";
-import { AZURE_LINKS } from "../../utils/constants/constants";
+import { AZURE_LINKS } from "../../utils/constants/azure";
 import { useSelector } from "react-redux";
 import { AppState } from "../../store/combineReducers";
 import { getSubscriptionsSelector } from "../../store/config/azure/selector";
