+++ src/client/src/containers/CosmosDbModal/index.tsx
@@ -9,7 +9,8 @@ import ApiSelection from "./APISelection/index";
 import SubscriptionSelection from "../../components/SubscriptionSelection";
 import { InjectedIntlProps, injectIntl } from "react-intl";
 import buttonStyles from "../../css/buttonStyles.module.css";
-import { KEY_EVENTS, AZURE, SERVICE_KEYS, AzureResourceType } from "../../utils/constants/constants";
+import { KEY_EVENTS } from "../../utils/constants/constants";
+import { AZURE, SERVICE_KEYS, AzureResourceType } from "../../utils/constants/azure";
 import styles from "./styles.module.css";
 import { AppState } from "../../store/combineReducers";
 import { ICosmosDB } from "../../store/userSelection/services/cosmosDb/model";
