+++ src/client/src/utils/constants/constants.ts
@@ -11,14 +11,6 @@ const WEB_TEMPLATE_STUDIO_LINKS = {
   ISSUES: "https://github.com/Microsoft/WebTemplateStudio/issues"
 };
 
-const AZURE_LINKS = {
-  CREATE_NEW_SUBSCRIPTION: "https://account.azure.com/signup?showCatalog=True&appId=SubscriptionsBlade",
-  CREATE_NEW_RESOURCE_GROUP: "https://portal.azure.com/#create/Microsoft.ResourceGroup",
-  APP_SERVICE_PLAN:"https://azure.microsoft.com/en-us/pricing/details/app-service/plans/",
-  VIEW_GENERATE_APP_SERVICE: "https://portal.azure.com/#blade/HubsExtension/BrowseResourceBlade/resourceType/Microsoft.Web%2Fsites",
-  VIEW_GENERATE_MONGO_DB: "https://portal.azure.com/#blade/HubsExtension/BrowseResourceBlade/resourceType/Microsoft.DocumentDb%2FdatabaseAccounts"
-};
-
 const PRODUCTION = "production";
 const DEVELOPMENT = "development";
 const TEST = "test";
@@ -58,20 +50,6 @@ const SERVICE_KEYS = {
   APP_SERVICE: "appService"
 };
 
-enum AzureResourceType {
-  AppService = "app-service",
-  Cosmos = "cosmos"
-}
-
-const AZURE = {
-  COSMOS_APIS: {
-    MONGO: "MongoDB",
-    SQL: "SQL"
-  },
-  DEFAULT_LOCATION: "Central US",
-  DEFAULT_RESOURCE_GROUP: ""
-}
-
 enum FRAMEWORK_TYPE {
   FRONTEND = "frontend",
   BACKEND = "backend"
@@ -81,11 +59,6 @@ enum KEY_EVENTS {
   ENTER = "Enter",
   SPACE = " ",
   TAB = "Tab"
-}
-
-enum SERVICE_GROUPS {
-  HOSTING = "CloudHosting",
-  DATABASE = "CloudDatabase",
 }
 
 const TELEMETRY = {
@@ -98,19 +71,15 @@ export {
   SERVICE_KEYS,
   INTL_MESSAGES,
   ARIA_LABELS_NAVIGATION,
-  AZURE,
   DEVELOPMENT,
   TEST,
   PROJECT_NAME_CHARACTER_LIMIT,
   PAGE_NAME_CHARACTER_LIMIT,
   MAX_PAGES_ALLOWED,
   WEB_TEMPLATE_STUDIO_LINKS,
-  AZURE_LINKS,
   FRAMEWORK_TYPE,
   KEY_EVENTS,
   PAYLOAD_MESSAGES_TEXT,
   BOOTSTRAP_LICENSE,
-  TELEMETRY,
-  AzureResourceType,
-  SERVICE_GROUPS
+  TELEMETRY
 };
